buy nft

+++ src/contracts/NFT.sol
@@ -6,7 +6,6 @@ import "@openzeppelin/contracts/token/ERC721/extensions/ERC721URIStorage.sol";
 import "@openzeppelin/contracts/utils/Counters.sol";
 
 contract NFT is ERC721URIStorage {
-
     using Counters for Counters.Counter;
     Counters.Counter private tokenIdss;
     address contractAddress;
@@ -15,7 +14,7 @@ contract NFT is ERC721URIStorage {
         contractAddress = marketplaceAddress;
     }
 
-    function mint(string memory _tokenURI) external returns (uint) {
+    function mint(string memory _tokenURI) external returns (uint256) {
         tokenIdss.increment();
         _safeMint(msg.sender, tokenIdss.current());
         _setTokenURI(tokenIdss.current(), _tokenURI);
